Use same versionless doc directory as official packages from Kitware
The doc directory used to be doc/cmake-<version>, but it was changed
at some point to the versionless doc/cmake instead. This was to prevent
browser bookmarks from having to be updated with each feature
release. The CMake snap should have also had this change at the same
time, but was overlooked until now.